perf: udpate

+++ packages/fx-core/src/component/driver/aad/utility/aadAppClient.ts
@@ -100,14 +100,16 @@ export class AadAppClient {
             this.is400Error(error), // sometimes AAD will complain OAuth permission not found if we pre-authorize a newly created permission
         },
       });
+      return response;
+    } catch (err) {
       if (
-        response.status === 400 &&
-        response.data.error.code === aadErrorCode.permissionErrorCode
+        axios.isAxiosError(err) &&
+        err.response &&
+        err.response.status === 400 &&
+        err.response.data.error.code === aadErrorCode.permissionErrorCode
       ) {
         throw new DeleteOrUpdatePermissionFailedError(AadAppClient.name);
       }
-      return response;
-    } catch (err) {
       throw err;
     }
   }
